test_integrationclient
